Finish firebase-storage integration

+++ RealEstatePropertyListingPlatform/RealEstatePropertyListingPlatform.API/Controllers/ImagesController.cs
@@ -1,7 +1,11 @@
 ﻿using System.Net;
+using System.Security.Claims;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Identity.Web;
 using RealEstatePropertyListingPlatform.Application.Contracts;
+using RealEstatePropertyListingPlatform.Application.Contracts.Interfaces;
+using WebAPI.Services;
 
 namespace RealEstatePropertyListingPlatform.API.Controllers
 {
@@ -12,13 +16,15 @@ namespace RealEstatePropertyListingPlatform.API.Controllers
 
     [ApiController]
     [Route("api/[controller]")]
-    public class UploadController : ControllerBase
+    public class ImagesController : ControllerBase
     {
         private readonly IImageStorageService _imageStorageService;
+        private readonly ICurrentUserService _currentUserService;
 
-        public UploadController(IImageStorageService imageStorageService)
+        public ImagesController(IImageStorageService imageStorageService, ICurrentUserService currentUserService)
         {
             _imageStorageService = imageStorageService;
+            _currentUserService = currentUserService;
         }
 
         [HttpGet("generate-upload-url")]
@@ -48,10 +54,20 @@ namespace RealEstatePropertyListingPlatform.API.Controllers
         }
 
         [HttpDelete("{fileName}")]
-        public async Task<IActionResult> Delete(string fileName)
+        public async Task<IActionResult> DeleteImage(string fileName)
         {
             try
             {
+                var currentUserId = _currentUserService.GetCurrentClaimsPrincipal().GetNameIdentifierId();
+                if (fileName == null)
+                {
+                    return BadRequest("Invalid file name");
+                }
+
+                if (fileName.Split('_').First() != currentUserId.ToString())
+                {
+                    return Forbid();
+                }
                 await _imageStorageService.DeleteImageAsync(fileName);
                 return NoContent();
             }
